requirements.txt
asgiref==3.5.2
Babel==2.10.3
colorama==0.4.5
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-cors-headers==3.13.0
django-import-export==2.9.0
django-jazzmin==2.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.5.0
pytz==2022.2.1
PyYAML==6.0
qrcode==7.3.1
reportbro-fpdf2==0.9.1
reportbro-lib==3.0.0
reportbro-simpleeval==0.9.11
sqlparse==0.4.2
svg.path==6.2
tablib==3.2.1
tzdata==2022.2
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
